+++ lib/arcanus/errors.rb
@@ -27,8 +27,9 @@ module Arcanus::Errors
   # Raised when invalid/non-existent command was used.
   class CommandInvalidError < UsageError; end
 
-  # Raised when run in a directory not part of a valid git repository.
-  class InvalidGitRepoError < UsageError; end
+  # Raised when run in a directory not part of a repository with Arcanus
+  # installed.
+  class InvalidArcanusRepoError < UsageError; end
 
   # Raised when a key path corresponding to a non-existent key is specified.
   class InvalidKeyPathError < UsageError; end