+++ lib/arcanus/repo.rb
@@ -12,44 +12,22 @@ module Arcanus
     # current working directory resides within.
     #
     # @return [String]
-    # @raise [Arcanus::Errors::InvalidGitRepoError] if the current directory
+    # @raise [Arcanus::Errors::InvalidArcanusRepoError] if the current directory
     #   doesn't reside within a git repository
     def root
       @root ||=
         begin
-          git_dir = Pathname.new(File.expand_path('.'))
+          arc_dir = Pathname.new(File.expand_path('.'))
                             .enum_for(:ascend)
                             .find do |path|
-            (path + '.git').exist?
+            (path + '.arcanus').exist?
           end
 
-          unless git_dir
-            raise Errors::InvalidGitRepoError, 'no .git directory found'
+          unless arc_dir
+            raise Errors::InvalidArcanusRepoError, 'no .arcanus directory found'
           end
 
-          git_dir.to_s
-        end
-    end
-
-    # Returns an absolute path to the .git directory for a repo.
-    #
-    # @return [String]
-    def git_dir
-      @git_dir ||=
-        begin
-          git_dir = File.expand_path('.git', root)
-
-          # .git could also be a file that contains the location of the git directory
-          unless File.directory?(git_dir)
-            git_dir = File.read(git_dir)[/^gitdir: (.*)$/, 1]
-
-            # Resolve relative paths
-            unless git_dir.start_with?('/')
-              git_dir = File.expand_path(git_dir, repo_dir)
-            end
-          end
-
-          git_dir
+          arc_dir.to_s
         end
     end