+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Arcanus Changelog
 
+## master (unreleased)
+
+* Don't require `.git` directory to exist
+
 ## 0.9.0
 
 * Pretty print hash in show command if key is a nested hash